Fix Containers and DAOs

+++ src/models/containers/firebase.container.ts
@@ -35,22 +35,26 @@ class FirebaseContainer {
 
   async getById(id) {
     const docRef = await this.collection.doc(id)
-    if (!docRef) {
+    const doc = await docRef.get()
+    if (!doc.exists) {
       const message = `Resource with id ${id} does not exists`
       throw new HttpError(HTTP_STATUS.NOT_FOUND, message)
     }
-    const document = await docRef.get()
-    return document.data()
+    return {
+      id: doc.id,
+      ...doc.data()
+    }
   }
 
-  async save(item) {
+  async save(item = { timestamp: new Date().toLocaleString(), products: [] }) {
     const docRef = this.collection.doc()
     return await docRef.set(item)
   }
 
   async update(id, item) {
-    const docRef = this.collection.doc(id)
-    if (!docRef) {
+    const docRef = await this.collection.doc(id)
+    const doc = await docRef.get()
+    if (!doc.exists) {
       const message = `Resource with id ${id} does not exists`
       throw new HttpError(HTTP_STATUS.NOT_FOUND, message)
     }
@@ -58,7 +62,12 @@ class FirebaseContainer {
   }
 
   async delete(id) {
-    const docRef = this.collection.doc(id)
+    const docRef = await this.collection.doc(id)
+    const doc = await docRef.get()
+    if (!doc.exists) {
+      const message = `Resource with id ${id} does not exists`
+      throw new HttpError(HTTP_STATUS.NOT_FOUND, message)
+    }
     return await docRef.delete()
   }
 }
